C driver name resolver (#886) (#890)
* [C] Add initial implementation of driver name resolver.  #886.

* [C] Ensure initialisation completes correctly.  #886.

* [C] Adjust naming convention.  #886.

* [C] Add close function to name resolver.  #886.

* [C] Add do_work function to name resolver.  #886.

* [C] Basic neighbor message passing.  #886.

* [C] Better message serialisation.  #886.

* [C] Basic name resolver cache.  #886.

* [C] Basic end to end sending of self resolutions.  #886.

* [C] Fix segfault in test.  #886.

* [C] Move values to fields for auto closing.  #886.

* [C] Basic gossip protocol working.  #886.

* [C] Fix compiler warnings.  #886.

* [C] Timeout entries in name resolver cache.  #886.

* [C] Re-order parameters.  #886.

* [C] Clean up test initialisation and add counter init.  #886.

* [C] Add neighbor counter.  #886.

* [C] Add cache size counter.  #886.

* [C] Timeout neighbors.  #886.

* [C] Group resolver functionality int tests.  #886.

* [C] Add distinct error logging to driver name resolver.  #886.

* [C] Re-ordering functions.  #886.

* [C] Align driver name resolver network buffer.  #886.

* [C] Cache string name length.  #886.

* [C] Handle IPv6 wildcards in driver name resolver.  #886.

* [C] Test tidy up.  #886.

* [C] Ensure that driver name resolver is properly wired in, including counters.  Update time_of_last_work correctly.  Some hacky logging.  #886.

* [C] Use datapaths correctly and start dissecting resolution frames.  #886.

* [C] Handle bootstrap sending correctly.  Re-resolve bootstrap neighbor periodically.  Handle self resolutions.  #886.

* [C] Implement log dissection for name resolution frames.  #886.

* [C] Add resolver properties to print configuration.  #886.

* [C] Various fixes for Mac and Windows.  #886.

* [C] Fix compile errors and warnings in MSVC build. #886

* [C] More error content on test failure.  #886.

* [C] Reset msghdr before every send.  Use correct msgname_len.  #886.

* [C] More error reporting and logging for Mac OS failures.  #886.

* [C] More error reporting and logging for Mac OS failures (take 2).  #886.

* [C] Allow time for message to propagate via local loopback.  #886.

* [C] Allow time for message to propagate via local loopback (one more test).  #886.

* [C] Add bytes_{sent,received} and short_sends counters to driver name resolver.  Report values in test on failure.  #886.

* [C] Get the counter variables correct.  #886.

* [C] Even more counters.  #886.

* [C] Move timestamp forward by 10ms each time we retry to ensure the duty cycle runs.  #886.

* [C] Sleep by 10ms when needing to retry duty cycle.  Tidy up use of counters in tests.  #886.

* [C] Renaming to aeron_default_name_resolver, aeron_driver_name_resolver, aeron_csv_table_name_resolver, and aeron_name_resolver_cache.  #886.

+++ aeron-client/src/main/c/concurrent/aeron_broadcast_receiver.c
@@ -141,7 +141,7 @@ int aeron_broadcast_receiver_receive(
             return -1;
         }
 
-        handler(type_id, (uint8_t *)receiver->scratch_buffer, 0, length);
+        handler(type_id, (uint8_t *)receiver->scratch_buffer, 0, (int)length);
 
         messages_received = 1;
     }
